use one source of truth for speechbrain data loader
