--- asr/test and presentation/speechbrain/asr-dataset-requirements.txt
@@ -1,1 +1,0 @@
-librosa>=0.8.1
