+++ asr/test and presentation/speechbrain/recipes/NIH/ASR_3/CTC/nih_prepare.py
@@ -1,0 +1,1 @@
+../../nih_prepare.py
