+++ asr/test and presentation/speechbrain/recipes/NIH/README.md
@@ -1,0 +1,1 @@
+See the ASR sub-directories for training instructions.
