+++ asr/test and presentation/speechbrain/requirements.txt
@@ -1,5 +1,4 @@
 -r lint-requirements.txt
--r asr-dataset-requirements.txt
 huggingface_hub>=0.0.6
 hyperpyyaml>=0.0.1
 joblib>=0.14.1
